+++ @utahdts/utah-design-system-header/package.json
@@ -71,7 +71,7 @@
     "@vitest/ui": "0.34.6",
     "jsdom": "22.1.0",
     "replace": "1.2.2",
-    "sass": "1.69.4",
+    "sass": "1.69.5",
     "tslib": "2.6.2",
     "typescript": "5.2.2",
     "vite": "4.5.0",
