+++ @utahdts/utah-design-system/package.json
@@ -69,12 +69,12 @@
     "uuid": "9.0.1"
   },
   "devDependencies": {
-    "@types/react": "18.2.29",
-    "@types/react-dom": "18.2.13",
+    "@types/react": "18.2.33",
+    "@types/react-dom": "18.2.14",
     "@vitejs/plugin-react": "4.1.0",
     "@vitest/coverage-istanbul": "0.34.6",
     "jsdom": "22.1.0",
-    "sass": "1.69.4",
+    "sass": "1.69.5",
     "vite": "4.5.0",
     "vitest": "0.34.6"
   },
