+++ examples/typed/typed-utah-header/package.json
@@ -15,14 +15,14 @@
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@types/react": "18.2.29",
-    "@types/react-dom": "18.2.13",
+    "@types/react": "18.2.33",
+    "@types/react-dom": "18.2.14",
     "@typescript-eslint/eslint-plugin": "6.7.0",
     "@typescript-eslint/parser": "6.7.0",
     "@vitejs/plugin-react": "4.1.0",
-    "eslint": "8.51.0",
+    "eslint": "8.52.0",
     "eslint-plugin-react-hooks": "4.6.0",
-    "eslint-plugin-react-refresh": "0.4.3",
+    "eslint-plugin-react-refresh": "0.4.4",
     "typescript": "5.2.2",
     "vite": "4.5.0"
   }
