+++ examples/utah-header/create-react-app/package.json
@@ -10,7 +10,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-scripts": "5.0.1",
-    "sass": "1.69.4",
+    "sass": "1.69.5",
     "web-vitals": "3.4.0"
   },
   "scripts": {
