+++ package.json
@@ -57,13 +57,13 @@
     "sitemap": "run-s generateSitemap submitSitemap"
   },
   "devDependencies": {
-    "eslint": "8.51.0",
+    "eslint": "8.52.0",
     "eslint-config-airbnb": "19.0.4",
-    "eslint-plugin-import": "2.28.1",
+    "eslint-plugin-import": "2.29.0",
     "eslint-plugin-jsx-a11y": "6.7.1",
     "eslint-plugin-react": "7.33.2",
     "eslint-plugin-react-hooks": "4.6.0",
-    "eslint-plugin-react-refresh": "0.4.3",
+    "eslint-plugin-react-refresh": "0.4.4",
     "npm-run-all": "4.1.5",
     "rimraf": "5.0.5",
     "vite-plugin-eslint": "1.8.1",
