+++ utah-design-system-website/package.json
@@ -28,11 +28,11 @@
     "use-immer": "0.9.0"
   },
   "devDependencies": {
-    "@types/react": "18.2.29",
-    "@types/react-dom": "18.2.13",
+    "@types/react": "18.2.33",
+    "@types/react-dom": "18.2.14",
     "@vitejs/plugin-react": "4.1.0",
     "jsdom": "22.1.0",
-    "sass": "1.69.4",
+    "sass": "1.69.5",
     "vite": "4.5.0",
     "vite-plugin-eslint": "1.8.1",
     "vite-plugin-mkcert": "1.16.0",
