+++ utah-design-system-website/src/react/components/preCode/PreCodeDefaultProps.jsx
@@ -1,0 +1,8 @@
+export const PreCodeDefaultProps = {
+  addHorizontalPadding: false,
+  allowScrollOverflow: false,
+  className: '',
+  maxHeight: null,
+  propsForPre: {},
+  showBackgroundColor: false,
+};
