+++ utah-design-system-website/src/react/components/preCode/PreCodeProps.jsx
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types';
+
+export const PreCodeProps = {
+  addHorizontalPadding: PropTypes.bool,
+  allowScrollOverflow: PropTypes.bool,
+  className: PropTypes.string,
+  maxHeight: PropTypes.string,
+  propsForPre: PropTypes.object,
+  showBackgroundColor: PropTypes.bool,
+};
